Makefile

+++ validation.c
@@ -6,7 +6,7 @@
 /*   By: eshor <eshor@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/07 15:53:40 by mbrogg            #+#    #+#             */
-/*   Updated: 2020/01/13 16:21:04 by eshor            ###   ########.fr       */
+/*   Updated: 2020/01/13 16:34:54 by eshor            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,7 +44,7 @@ char	check_of_ending_block(t_lst **head, short **mas, int *i, char *ar)
 	y = 0;
 	if (++(*i) == 4)
 	{
-		if (!check_area((*head)->coords, x, y))
+		if (!check_area((*head)->coords, (int)x, (int)y))
 			return (0);
 		ar[0] = 0;
 		ar[1]++;
